// src/generated/api/customers/customers.schemas.ts
export type CustomerStatus = (typeof CustomerStatus)[keyof typeof CustomerStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CustomerStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;

export type CustomerKind = (typeof CustomerKind)[keyof typeof CustomerKind];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CustomerKind = {
  MEMBER: 'MEMBER',
  GUEST: 'GUEST',
} as const;